requirements.txt
Django==1.8.2
django-rest-swagger==0.3.2
djangorestframework==3.1.3
flake8==2.4.1
gnureadline==6.3.3
httpie==0.9.2
ipython==3.1.0
mccabe==0.3
pep8==1.5.7
pyflakes==0.8.1
Pygments==2.0.2
PyYAML==3.11
requests==2.7.0
wheel==0.24.0
